% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_loss_functions.R
\name{custom.xgboost.sera}
\alias{custom.xgboost.sera}
\title{Custom SERA loss function adaptation for XGBoost.}
\usage{
custom.xgboost.sera(preds, dtrain)
}
\arguments{
\item{preds}{Numeric \code{vector} of predictions.}

\item{dtrain}{\code{xgb.DMatrix} object.}
}
\value{
Returns a \code{list with the following elements:
}
\description{
Determines the gradient statistics of SERA to be used in
XGBoost algorithm as a custom loss function.
}
\examples{

}
\references{
https://github.com/dmlc/xgboost/blob/master/R-package/demo/custom_objective.R
}
