% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_extreme_type.R
\name{get_extreme_type}
\alias{get_extreme_type}
\title{Get extreme type.}
\usage{
get_extreme_type(y)
}
\arguments{
\item{y}{numeric \code{vector}.}
}
\value{
One of type of extreme: "high", "low", "both".
}
\description{
Returns the type of an extreme depending on the existence of
outliers in the adjusted boxplot.
}
\examples{
library(robustbase)

y <- rnorm(100)
extreme_type <- get_extreme_type(y)
}
