% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GradientTreeBoosting.R
\name{GradientTreeBoost}
\alias{GradientTreeBoost}
\title{Gradient Tree Boosting}
\usage{
GradientTreeBoost(formula, train, test, maxIter = 100, eta = 0.01, verbose = 0)
}
\arguments{
\item{formula}{A \code{formula object}.}

\item{train}{A \code{data.frame} or \code{tibble} object with the training set.}

\item{test}{A \code{data.frame} or \code{tibble} object with the test set.}

\item{maxIter}{The maximum number of iterations.}

\item{eta}{Learning rate.}

\item{verbose}{Prints out the error across iterations (if 1)}
}
\value{
A numeric vector with predictions and execution time (in seconds).
}
\description{
Determines predictions of a given dataset using the Gradient Tree Boost for Regression using
MSE as an optimisation loss function.
}
\examples{
\dontrun{

library(dplyr)
library(rpart)
library(treeClust)

data(NO2Emissions)

n <- nrow(NO2Emissions)
s <- sample(1:n, size = n*0.8)

formula <- LNO2 ~ .
train <- NO2Emissions \%>\% dplyr::slice(s)
test <- NO2Emissions \%>\% dplyr::slice(-s)

res <- GradientTreeBoost(formula, train, test)
res
}
}
