% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_loss_functions.R
\name{lgbm.sera}
\alias{lgbm.sera}
\title{Custom SERA loss function adaptation for Light GBM.}
\usage{
lgbm.sera(preds, dtrain)
}
\arguments{
\item{preds}{Numeric \code{vector} of predictions.}

\item{dtrain}{A \code{lgb.Dataset} object.}
}
\value{
Returns a list containing the gradient and the hessian of SERA.
}
\description{
Determines the gradient statistics of SERA to be used in
LGBM algorithm as a custom loss function.
}
\references{
https://github.com/microsoft/LightGBM/blob/master/R-package/tests/testthat/test_custom_objective.R
}
