% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NO2Emissions}
\alias{NO2Emissions}
\title{Log of NO2 Emissions in air.}
\format{
A data frame with 500 observations and 8 variables.
}
\usage{
data(NO2Emissions)
}
\description{
Log of NO2 Emissions in air.
}
\keyword{datasets}
