% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sera_derivatives.R
\name{sera_first_deriv}
\alias{sera_first_deriv}
\title{SERA first derivative}
\usage{
sera_first_deriv(trues, preds, phi, step = 0.001)
}
\arguments{
\item{trues}{Target values from a test set of a given data set. Should be a vector and have the same size as the variable preds}

\item{preds}{Predicted values given a certain test set of a given data set. Should be a vector and have the same size}

\item{phi}{Relevance of the values in the parameter trues.}

\item{step}{Relevance intervals between 0 (min) and 1 (max). Default 0.001.}
}
\value{
A vector with SERA first derivative for each prediction.
}
\description{
Calculates SERA first derivative using the trapezoidal rule.
}
\examples{
\dontrun{
library(dplyr)
library(rpart)
library(IRon)

data("NO2Emissions")

n <- nrow(NO2Emissions)
s <- sample(1:n, size = n*0.8)

formula <- LNO2 ~ .
train <- NO2Emissions \%>\% slice(s)
test <- NO2Emissions \%>\% slice(-s)

trues <- train$LNO2
phi <- phi(trues)

model <- rpart(formula, train)
preds <- predict(model, test)

sera_deriv <- sera_first_deriv(trues, preds, phi)
}
}
