% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sera_derivatives.R
\name{sera_first_deriv_num}
\alias{sera_first_deriv_num}
\title{SERA numerical first derivative based on sigma.}
\usage{
sera_first_deriv_num(trues, preds, sigmas)
}
\arguments{
\item{trues}{A vector with the true values of a variable.}

\item{preds}{A vector with the prediction values of a model.}

\item{sigmas}{A vector with the sigmas returned from calling \code{\link{sigma}}.}
}
\value{
A vector with SERA first derivative for each prediction.
}
\description{
SERA numerical first derivative based on sigma.
}
\examples{
\dontrun{
library(dplyr)
library(rpart)
library(IRon)

data("NO2Emissions")

n <- nrow(NO2Emissions)
s <- sample(1:n, size = n*0.8)

formula <- LNO2 ~ .
train <- NO2Emissions \%>\% slice(s)
test <- NO2Emissions \%>\% slice(-s)

trues <- train$LNO2
phi <- phi(trues)
steps <- seq(0, 1, 0.001)
sigmas <- sigma(phi, steps)

model <- rpart(formula, train)
preds <- predict(model, test)

sera_first_deriv <- sera_first_deriv_num(trues, preds, sigmas)
}
}
