% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sera_derivatives.R
\name{sera_min}
\alias{sera_min}
\title{Calculates the value for which SERA is minimum.}
\usage{
sera_min(trues, phi.trues = NULL, ph = NULL, step = 0.001)
}
\arguments{
\item{trues}{A vector of true values.}

\item{phi.trues}{A vector of the relevance function w.r.t. true values.}

\item{ph}{Control points.}

\item{step}{The step used in the trapezoidal rule. Defaults to 0.001.}
}
\value{
A numeric value for which SERA is minimum.
}
\description{
Calculates the value for which SERA is minimum.
}
\examples{
\dontrun{
library(dplyr)
library(IRon)

data("NO2Emissions")

n <- nrow(NO2Emissions)
s <- sample(1:n, size = n*0.8)

formula <- LNO2 ~ .
train <- NO2Emissions \%>\% slice(s)
test <- NO2Emissions \%>\% slice(-s)

trues <- train$LNO2
phi <- phi(trues)

sera_min <- sera_min(trues, phi)
}
}
