% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sera.R
\name{sera_numerical}
\alias{sera_numerical}
\title{SERA numerical approximation based on sigma.}
\usage{
sera_numerical(trues, preds, sigmas)
}
\arguments{
\item{trues}{A vector with the true values of a variable.}

\item{preds}{A vector with the prediction values of a model.}

\item{sigmas}{A vector with the sigmas returned from calling \code{\link{sigma}}.}
}
\value{
A scalar with SERA error.
}
\description{
SERA numerical approximation based on sigma.
}
\examples{

library(dplyr)
library(rpart)
library(IRon)

data("NO2Emissions")

n <- nrow(NO2Emissions)
s <- sample(1:n, size = n*0.8)

formula <- LNO2 ~ .
train <- NO2Emissions \%>\% slice(s)
test <- NO2Emissions \%>\% slice(-s)

trues <- train$LNO2
phi <- phi(trues)
steps <- seq(0, 1, 0.001)
sigmas <- sigma(phi, steps)

model <- rpart(formula, train)
preds <- predict(model, test)

sera <- sera_numerical(trues, preds, sigmas)


}
