% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sera_derivatives.R
\name{sera_second_deriv}
\alias{sera_second_deriv}
\title{SERA second derivative}
\usage{
sera_second_deriv(trues, phi, step = 0.001)
}
\arguments{
\item{trues}{Target values from a test set of a given data set. Should be a vector and have the same size as the variable preds}

\item{phi}{Relevance of the values in the parameter trues.}

\item{step}{Relevance intervals between 0 (min) and 1 (max). Default 0.001.}
}
\value{
A vector with SERA second derivative for each prediction.
}
\description{
Calculates SERA second derivative using the trapezoidal rule with Riemann's sum.
}
\examples{

library(dplyr)
library(IRon)

data("NO2Emissions")

n <- nrow(NO2Emissions)
s <- sample(1:n, size = n*0.8)

formula <- LNO2 ~ .
train <- NO2Emissions \%>\% slice(s)
test <- NO2Emissions \%>\% slice(-s)

trues <- train$LNO2
phi <- phi(trues)

sera_second_deriv <- sera_second_deriv(trues, phi)

}
