% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sera_derivatives.R
\name{sera_second_deriv_num}
\alias{sera_second_deriv_num}
\title{SERA numerical second derivative based on sigma.}
\usage{
sera_second_deriv_num(sigmas)
}
\arguments{
\item{sigmas}{A vector with the sigmas returned from calling \code{\link{sigma}}.}
}
\value{
A vector with SERA second derivative for each prediction.
}
\description{
SERA numerical second derivative based on sigma.
}
\examples{
\dontrun{
library(dplyr)
library(IRon)

data("NO2Emissions")

n <- nrow(NO2Emissions)
s <- sample(1:n, size = n*0.8)

formula <- LNO2 ~ .
train <- NO2Emissions \%>\% slice(s)
test <- NO2Emissions \%>\% slice(-s)

trues <- train$LNO2
phi <- phi(trues)
steps <- seq(0, 1, 0.001)
sigmas <- sigma(phi, steps)

sera_second_deriv <- sera_second_deriv_num(sigmas)
}
}
