% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma.R
\name{sigma}
\alias{sigma}
\title{Calculates the number of steps up to a given relevance phi, for all phis.}
\usage{
sigma(phis, steps)
}
\arguments{
\item{phis}{A vector with the relevance of the target variable.}

\item{steps}{A numeric vector with steps.}
}
\value{
A numeric vector with sigma values.
}
\description{
Calculates the number of steps up to a given relevance phi, for all phis.
}
\examples{

\dontrun{

library(IRon)

y <- rnorm(100)
phi <- phi(y)
steps <- seq(0, 1, 0.001)
sigmas <- sigma(phi, steps)

sigmas <- sigma(phi, steps)
}
}
