% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma.R
\name{sigma_one}
\alias{sigma_one}
\title{Get sigma.}
\usage{
sigma_one(phi, steps)
}
\arguments{
\item{phi}{A number with the relevance of the target variable.}

\item{steps}{A numeric vector with steps.}
}
\value{
A number with the number of occurrences of a given observation.
}
\description{
Calculates the number of steps up to a given relevance phi.
}
\examples{
\dontrun{

library(IRon)

y <- rnorm(100)
phis <- phi(y)
steps <- seq(0, 1, 0.001)

phi <- phis[1]
sigmas <- sigma_one(phi, steps)
sigmas <- sigma_one(phi, steps)
}
}
