% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{wf.SERAGradientTreeBoost}
\alias{wf.SERAGradientTreeBoost}
\title{Gradient Boosting Regression Trees (SERA)}
\usage{
wf.SERAGradientTreeBoost(formula, train, test, ...)
}
\arguments{
\item{formula}{A \code{formula} object.}

\item{train}{\code{data.frame} or \code{tibble} object with the training set.}

\item{test}{\code{data.frame} or \code{tibble} object with the test set.}

\item{...}{Additional parameters which can be passed into internal model.}
}
\value{
A \code{list} containing trues, predictions and execution time (in seconds).
}
\description{
Workflow for Gradient Boosting Regression Trees optimised with SERA.
}
\examples{
\dontrun{

library(dplyr)
library(IRon)
library(rpart)
library(treeClust)

data("NO2Emissions")

n <- nrow(NO2Emissions)
s <- sample(1:n, size = n*0.8)

formula <- LNO2 ~ .
train <- NO2Emissions \%>\% dplyr::slice(s)
test <- NO2Emissions \%>\% dplyr::slice(-s)

res <- wf.SERAGradientTreeBoost(formula, train, test)
res
}
}
