% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{wf.XGBoost_SERA}
\alias{wf.XGBoost_SERA}
\title{XGBoost(SERA) workflow}
\usage{
wf.XGBoost_SERA(formula, train, test, nrounds = 100, ...)
}
\arguments{
\item{formula}{A \code{formula} object.}

\item{train}{\code{data.frame} or \code{tibble} object with the training set.}

\item{test}{\code{data.frame} or \code{tibble} object with the test set.}

\item{nrounds}{A number. Number of boosters to train.}

\item{...}{Additional parameters which can be passed into internal model.}
}
\value{
A \code{list} containing trues, predictions and execution time (in seconds).
}
\description{
Workflow for XGBoost optimised with SERA.
}
\details{
Calls \code{\link{custom.xgboost.sera}} as \code{objective}.
}
\examples{
\dontrun{

library(dplyr)
library(IRon)
library(xgboost)
library(robustbase)

data("NO2Emissions")

n <- nrow(NO2Emissions)
s <- sample(1:n, size = n*0.8)

formula <- LNO2 ~ .
train <- NO2Emissions \%>\% dplyr::slice(s)
test <- NO2Emissions \%>\% dplyr::slice(-s)

nrounds = 100


res <- wf.XGBoost_SERA(formula, train, test, nrounds=nrounds)
res
}
}
