% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_loss_functions.R
\name{xgboost.sera}
\alias{xgboost.sera}
\title{Custom SERA loss function adaptation for XGBoost.}
\usage{
xgboost.sera(preds, dtrain)
}
\arguments{
\item{preds}{Numeric \code{vector} of predictions.}

\item{dtrain}{\code{xgb.DMatrix} object.}
}
\value{
Returns a list containing the gradient and the hessian of SERA.
}
\description{
Determines the gradient statistics of SERA to be used in
XGBoost algorithm as a custom loss function.
}
\examples{

}
\references{
https://github.com/dmlc/xgboost/blob/master/R-package/demo/custom_objective.R
}
